improve logic finalize

+++ contracts/MowjowToken.sol
@@ -8,7 +8,7 @@ contract MowjowToken is MintableToken {
     string public name = "MowjowToken";
     string public symbol = "MJT";
     uint public decimals = 18;
-    uint public INITIAL_SUPPLY = 75 * 1e8;
+    uint256 public INITIAL_SUPPLY = 75 * 1e8;
 
     function MowjowToken() {
         totalSupply = INITIAL_SUPPLY;
